POP movie project stage 2
POP movie project stage 2 - add more information to your movie details
view:

allow users to view and play trailers ( either in the youtube app or a
web browser).
Allow users to read reviews of a selected m ovie.Allow users to mark a
movie as a favorite in the details view by tapping a button(star). This
is for a local movies collection that you will maintain and does not
require an API request*.
Modify the existing sorting criteria for the main view to include an
additional pivot to show their favorites collection.
optimize app experience for tablet.